testing all delete and added warinng

+++ Project/DataModels/BoughtSnacksModel.cs
@@ -5,12 +5,12 @@ public class BoughtSnacksModel
     public int SnackId { get; set; }
     public int Amount { get; set; }
 
-    public BoughtSnacksModel(Int64 id, Int64 Reservation_ID, Int64 Snack_ID, int amount)
+    public BoughtSnacksModel(Int64 id, Int64 Reservation_ID, Int64 Snack_ID, Int64 amount)
     {
         Id = (int)id;
         ReservationId = (int)Reservation_ID;
         SnackId = (int)Snack_ID;
-        Amount = amount;
+        Amount = (int)amount;
     }
     public BoughtSnacksModel(int reservation_id, int snack_id, int amount)
     {
